requirements.txt
langchain_community
nx-arangodb
langchain-google-genai
google-genai
pydantic
fuzzywuzzy
python-Levenshtein
python-dotenv